split entity service
